+++ src/lib/components/flight-filters/MobileFiltersModal.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
-  import { CalendarDate, type DateValue } from '@internationalized/date';
   import {
     Calendar as CalendarIcon,
     Check,
-    ChevronLeft,
     ChevronRight,
     Hash,
     Plane,
@@ -15,12 +13,28 @@
     Plus,
   } from '@o7/icon';
   import { Airlines } from '@o7/icon/material';
-  import { DateRangePicker } from 'bits-ui';
   import type { Component } from 'svelte';
 
+  import MobileDateFilterScreen from './MobileDateFilterScreen.svelte';
   import UserAvatar from '$lib/components/display/UserAvatar.svelte';
   import {
-    defaultFilters,
+    clearFilterColumn,
+    createDefaultFilters,
+    dateFilterSummary,
+    isFilterColumnActive,
+    optionColumnOperator,
+    optionColumnValues,
+    pluralMultiOptionOperators,
+    pluralOptionOperators,
+    setOptionColumnOperator,
+    setOptionColumnValues,
+    singularMultiOptionOperators,
+    singularOptionOperators,
+    type FilterColumnId,
+    type NonPassengerOptionColumnId,
+    type OptionColumnId,
+  } from '$lib/components/flight-filters/model';
+  import {
     hasTempFilters as hasActiveTempFilters,
     type FlightFilters,
     type MultiOptionFilterOperator,
@@ -29,7 +43,6 @@
   } from '$lib/components/flight-filters/types';
   import AnimatedSizeContainer from '$lib/components/ui/animated-size-container.svelte';
   import { Button } from '$lib/components/ui/button';
-  import * as CalendarParts from '$lib/components/ui/calendar';
   import {
     Modal,
     ModalBody,
@@ -43,16 +56,6 @@
     type FlightData,
   } from '$lib/utils';
 
-  type FilterColumnId =
-    | 'departureAirports'
-    | 'arrivalAirports'
-    | 'date'
-    | 'passengers'
-    | 'airline'
-    | 'aircraft'
-    | 'aircraftRegs';
-
-  type OptionColumnId = Exclude<FilterColumnId, 'date'>;
   type Screen =
     | { kind: 'home' }
     | { kind: 'subjects' }
@@ -78,44 +81,11 @@
     username?: string;
   };
 
-  const optionOperators: Array<{ value: OptionFilterOperator; label: string }> =
-    [
-      { value: 'is', label: 'Is' },
-      { value: 'is not', label: 'Is not' },
-      { value: 'is any of', label: 'Is any of' },
-      { value: 'is none of', label: 'Is none of' },
-    ];
-  const singularOptionOperators = optionOperators.filter(
-    (operator) => operator.value === 'is' || operator.value === 'is not',
-  );
-  const pluralOptionOperators = optionOperators.filter(
-    (operator) =>
-      operator.value === 'is any of' || operator.value === 'is none of',
-  );
-
-  const multiOptionOperators: Array<{
-    value: MultiOptionFilterOperator;
-    label: string;
-  }> = [
-    { value: 'include', label: 'Include' },
-    { value: 'exclude', label: 'Exclude' },
-    { value: 'include any of', label: 'Include any of' },
-    { value: 'include all of', label: 'Include all of' },
-    { value: 'exclude if any of', label: 'Exclude if any of' },
-    { value: 'exclude if all', label: 'Exclude if all' },
-  ];
-  const singularMultiOptionOperators = multiOptionOperators.filter(
-    (operator) => operator.value === 'include' || operator.value === 'exclude',
-  );
-  const pluralMultiOptionOperators = multiOptionOperators.filter(
-    (operator) => operator.value !== 'include' && operator.value !== 'exclude',
-  );
-
   let {
     open = $bindable(),
     flights,
     filters = $bindable(),
-    tempFilters,
+    tempFilters = $bindable(),
     hasTempFilters = false,
   }: {
     open: boolean;
@@ -128,8 +98,6 @@
   let screen = $state<Screen>({ kind: 'home' });
   let search = $state('');
   let wasOpen = $state(open);
-  let calendarPlaceholder = $state<DateValue | undefined>(undefined);
-  let calendarMode = $state<'normal' | 'year' | 'month'>('normal');
 
   const tempLocationFiltersActive = $derived(
     hasTempFilters || hasActiveTempFilters(tempFilters),
@@ -163,22 +131,6 @@
   ]);
 
   const visibleColumns = $derived(columns.filter((column) => !column.hidden));
-
-  function cloneFilters(next: Partial<FlightFilters> = {}) {
-    filters = {
-      ...filters,
-      departureAirports: [...filters.departureAirports],
-      arrivalAirports: [...filters.arrivalAirports],
-      airportsEither: [...filters.airportsEither],
-      routes: [...filters.routes],
-      years: [...filters.years],
-      passengers: [...filters.passengers],
-      airline: [...filters.airline],
-      aircraft: [...filters.aircraft],
-      aircraftRegs: [...filters.aircraftRegs],
-      ...next,
-    };
-  }
 
   function resetNavigation() {
     screen = { kind: 'home' };
@@ -310,166 +262,38 @@
   });
 
   function valuesFor(columnId: OptionColumnId) {
-    return filters[columnId];
+    return optionColumnValues(filters, columnId);
   }
 
   function operatorFor(columnId: OptionColumnId) {
-    switch (columnId) {
-      case 'passengers':
-        return normalizeMultiOptionOperator(
-          filters.passengersOperator,
-          valuesFor(columnId).length,
-        );
-      case 'departureAirports':
-        return normalizeOptionOperator(
-          filters.departureAirportsOperator,
-          valuesFor(columnId).length,
-        );
-      case 'arrivalAirports':
-        return normalizeOptionOperator(
-          filters.arrivalAirportsOperator,
-          valuesFor(columnId).length,
-        );
-      case 'airline':
-        return normalizeOptionOperator(
-          filters.airlineOperator,
-          valuesFor(columnId).length,
-        );
-      case 'aircraft':
-        return normalizeOptionOperator(
-          filters.aircraftOperator,
-          valuesFor(columnId).length,
-        );
-      case 'aircraftRegs':
-        return normalizeOptionOperator(
-          filters.aircraftRegsOperator,
-          valuesFor(columnId).length,
-        );
+    if (columnId === 'passengers') {
+      return optionColumnOperator(filters, columnId);
     }
+    return optionColumnOperator(filters, columnId);
   }
 
-  function rawOperatorFor(columnId: OptionColumnId) {
-    switch (columnId) {
-      case 'passengers':
-        return filters.passengersOperator;
-      case 'departureAirports':
-        return filters.departureAirportsOperator;
-      case 'arrivalAirports':
-        return filters.arrivalAirportsOperator;
-      case 'airline':
-        return filters.airlineOperator;
-      case 'aircraft':
-        return filters.aircraftOperator;
-      case 'aircraftRegs':
-        return filters.aircraftRegsOperator;
-    }
-  }
-
-  function normalizeOptionOperator(
-    operator: OptionFilterOperator,
-    valueCount: number,
-  ): OptionFilterOperator {
-    if (valueCount > 1) {
-      if (operator === 'is') return 'is any of';
-      if (operator === 'is not') return 'is none of';
-      return operator;
-    }
-
-    if (operator === 'is any of') return 'is';
-    if (operator === 'is none of') return 'is not';
-    return operator;
-  }
-
-  function normalizeMultiOptionOperator(
-    operator: MultiOptionFilterOperator,
-    valueCount: number,
-  ): MultiOptionFilterOperator {
-    if (valueCount > 1) {
-      if (operator === 'include') return 'include any of';
-      if (operator === 'exclude') return 'exclude if any of';
-      return operator;
-    }
-
-    if (operator === 'exclude if any of' || operator === 'exclude if all') {
-      return 'exclude';
-    }
-
-    if (operator === 'include any of' || operator === 'include all of') {
-      return 'include';
-    }
-
-    return operator;
-  }
-
-  function normalizeOperatorForCount(
+  function setOperator(
     columnId: OptionColumnId,
-    valueCount = valuesFor(columnId).length,
+    operator: OptionFilterOperator | MultiOptionFilterOperator,
   ) {
-    const operator = rawOperatorFor(columnId);
-    return columnId === 'passengers'
-      ? normalizeMultiOptionOperator(
-          operator as MultiOptionFilterOperator,
-          valueCount,
-        )
-      : normalizeOptionOperator(operator as OptionFilterOperator, valueCount);
-  }
-
-  function setOperator(columnId: OptionColumnId, operator: string) {
-    switch (columnId) {
-      case 'passengers':
-        cloneFilters({
-          passengersOperator: operator as MultiOptionFilterOperator,
-        });
-        break;
-      case 'departureAirports':
-        cloneFilters({
-          departureAirportsOperator: operator as OptionFilterOperator,
-        });
-        break;
-      case 'arrivalAirports':
-        cloneFilters({
-          arrivalAirportsOperator: operator as OptionFilterOperator,
-        });
-        break;
-      case 'airline':
-        cloneFilters({ airlineOperator: operator as OptionFilterOperator });
-        break;
-      case 'aircraft':
-        cloneFilters({ aircraftOperator: operator as OptionFilterOperator });
-        break;
-      case 'aircraftRegs':
-        cloneFilters({
-          aircraftRegsOperator: operator as OptionFilterOperator,
-        });
-        break;
+    if (columnId === 'passengers') {
+      filters = setOptionColumnOperator(
+        filters,
+        columnId,
+        operator as MultiOptionFilterOperator,
+      );
+      return;
     }
-  }
 
-  function operatorPatch(columnId: OptionColumnId, operator: string) {
-    switch (columnId) {
-      case 'passengers':
-        return { passengersOperator: operator as MultiOptionFilterOperator };
-      case 'departureAirports':
-        return { departureAirportsOperator: operator as OptionFilterOperator };
-      case 'arrivalAirports':
-        return { arrivalAirportsOperator: operator as OptionFilterOperator };
-      case 'airline':
-        return { airlineOperator: operator as OptionFilterOperator };
-      case 'aircraft':
-        return { aircraftOperator: operator as OptionFilterOperator };
-      case 'aircraftRegs':
-        return { aircraftRegsOperator: operator as OptionFilterOperator };
-    }
+    filters = setOptionColumnOperator(
+      filters,
+      columnId as NonPassengerOptionColumnId,
+      operator as OptionFilterOperator,
+    );
   }
 
   function setValues(columnId: OptionColumnId, values: string[]) {
-    cloneFilters({
-      [columnId]: values,
-      ...operatorPatch(
-        columnId,
-        normalizeOperatorForCount(columnId, values.length),
-      ),
-    } as Partial<FlightFilters>);
+    filters = setOptionColumnValues(filters, columnId, values);
   }
 
   function toggleValue(columnId: OptionColumnId, value: string) {
@@ -481,20 +305,11 @@
   }
 
   function clearColumn(columnId: FilterColumnId) {
-    switch (columnId) {
-      case 'date':
-        cloneFilters({ years: [], fromDate: undefined, toDate: undefined });
-        break;
-      default:
-        setValues(columnId, []);
-        break;
-    }
+    filters = clearFilterColumn(filters, columnId);
   }
 
   function isColumnActive(columnId: FilterColumnId) {
-    if (columnId === 'date')
-      return !!(filters.years.length || filters.fromDate || filters.toDate);
-    return valuesFor(columnId).length > 0;
+    return isFilterColumnActive(filters, columnId);
   }
 
   const activeColumns = $derived(
@@ -515,118 +330,8 @@
     wasOpen = open;
   });
 
-  function formatDateLabel(value: CalendarDate | undefined) {
-    if (!value) return '';
-    return new Intl.DateTimeFormat(undefined, {
-      month: 'short',
-      day: 'numeric',
-      year: 'numeric',
-    }).format(new Date(value.year, value.month - 1, value.day));
-  }
-
   function dateSummary() {
-    if (filters.years.length) {
-      if (filters.years.length <= 2) return filters.years.join(', ');
-      return `${filters.years.slice(0, 2).join(', ')} +${filters.years.length - 2}`;
-    }
-
-    if (filters.fromDate && filters.toDate) {
-      if (filters.fromDate.compare(filters.toDate) === 0) {
-        return formatDateLabel(filters.fromDate);
-      }
-      return `${formatDateLabel(filters.fromDate)} - ${formatDateLabel(filters.toDate)}`;
-    }
-    if (filters.fromDate) return `From ${formatDateLabel(filters.fromDate)}`;
-    if (filters.toDate) return `Until ${formatDateLabel(filters.toDate)}`;
-    return 'Choose date';
-  }
-
-  function setDateMode(mode: 'single' | 'range' | 'from' | 'to') {
-    const today = new CalendarDate(
-      new Date().getFullYear(),
-      new Date().getMonth() + 1,
-      new Date().getDate(),
-    );
-    const anchor = filters.fromDate ?? filters.toDate ?? today;
-
-    if (mode === 'single') {
-      cloneFilters({ years: [], fromDate: anchor, toDate: anchor });
-    } else if (mode === 'range') {
-      cloneFilters({
-        years: [],
-        fromDate: filters.fromDate ?? anchor,
-        toDate: filters.toDate ?? anchor,
-      });
-    } else if (mode === 'from') {
-      cloneFilters({ years: [], fromDate: anchor, toDate: undefined });
-    } else {
-      cloneFilters({ years: [], fromDate: undefined, toDate: anchor });
-    }
-  }
-
-  function dateMode() {
-    if (filters.fromDate && filters.toDate) {
-      return filters.fromDate.compare(filters.toDate) === 0
-        ? 'single'
-        : 'range';
-    }
-    if (filters.fromDate) return 'from';
-    if (filters.toDate) return 'to';
-    return 'single';
-  }
-
-  function setThisYear() {
-    const year = new Date().getFullYear();
-    cloneFilters({
-      years: [],
-      fromDate: new CalendarDate(year, 1, 1),
-      toDate: new CalendarDate(year, 12, 31),
-    });
-  }
-
-  const dateRangeValue = $derived.by(() => ({
-    start: filters.fromDate,
-    end:
-      filters.fromDate && filters.toDate?.compare(filters.fromDate) === 0
-        ? undefined
-        : filters.toDate,
-  }));
-
-  $effect(() => {
-    const anchor = filters.fromDate ?? filters.toDate;
-    if (anchor && !calendarPlaceholder) calendarPlaceholder = anchor;
-  });
-
-  function handleDateRangeChange(
-    value:
-      | { start: DateValue | undefined; end: DateValue | undefined }
-      | undefined,
-  ) {
-    const start = value?.start as CalendarDate | undefined;
-    const end = value?.end as CalendarDate | undefined;
-    const mode = dateMode();
-
-    if (!start && !end) {
-      cloneFilters({ years: [], fromDate: undefined, toDate: undefined });
-      return;
-    }
-
-    const anchor = start ?? end;
-    if (anchor) calendarPlaceholder = anchor;
-
-    if (mode === 'single') {
-      cloneFilters({ years: [], fromDate: anchor, toDate: anchor });
-    } else if (mode === 'from') {
-      cloneFilters({ years: [], fromDate: anchor, toDate: undefined });
-    } else if (mode === 'to') {
-      cloneFilters({ years: [], fromDate: undefined, toDate: anchor });
-    } else {
-      cloneFilters({
-        years: [],
-        fromDate: start,
-        toDate: end ?? start,
-      });
-    }
+    return dateFilterSummary(filters);
   }
 
   function selectedOptions(columnId: OptionColumnId) {
@@ -656,11 +361,6 @@
     if (!labels.length) return 'Choose values';
     if (labels.length <= 2) return labels.join(', ');
     return `${labels.slice(0, 2).join(', ')} +${labels.length - 2}`;
-  }
-
-  function activeFilterSummary(column: FilterColumn) {
-    if (column.id === 'date') return dateSummary();
-    return `${operatorFor(column.id)} ${valueSummary(column.id)}`;
   }
 
   function filteredOptions(columnId: OptionColumnId) {
@@ -698,18 +398,7 @@
   }
 
   function clearAll() {
-    filters = {
-      ...defaultFilters,
-      departureAirports: [],
-      arrivalAirports: [],
-      airportsEither: [],
-      routes: [],
-      years: [],
-      passengers: [],
-      airline: [],
-      aircraft: [],
-      aircraftRegs: [],
-    };
+    filters = createDefaultFilters();
   }
 
   const title = $derived.by(() => {
@@ -928,21 +617,22 @@
           </div>
         </div>
       {:else if screen.kind === 'filter'}
-        {@const column = columnById(screen.columnId)}
+        {@const columnId = screen.columnId}
+        {@const column = columnById(columnId)}
         {#if column}
           <div class="flex max-h-[72dvh] flex-col">
             <div class="space-y-3 border-b p-3">
               <div class="flex gap-2 overflow-x-auto pb-0.5">
-                {#each operatorsFor(screen.columnId) as operator (operator.value)}
+                {#each operatorsFor(columnId) as operator (operator.value)}
                   <button
                     type="button"
                     class={cn(
                       'h-8 shrink-0 rounded-full border px-3 text-xs font-semibold transition-colors',
-                      operatorFor(screen.columnId) === operator.value
+                      operatorFor(columnId) === operator.value
                         ? 'border-foreground bg-foreground text-background'
                         : 'bg-background text-muted-foreground active:bg-muted',
                     )}
-                    onclick={() => setOperator(screen.columnId, operator.value)}
+                    onclick={() => setOperator(columnId, operator.value)}
                   >
                     {operator.label}
                   </button>
@@ -960,14 +650,12 @@
               </label>
             </div>
             <div class="min-h-0 flex-1 overflow-y-auto p-2">
-              {#each filteredOptions(screen.columnId) as option (option.value)}
-                {@const selected = valuesFor(screen.columnId).includes(
-                  option.value,
-                )}
+              {#each filteredOptions(columnId) as option (option.value)}
+                {@const selected = valuesFor(columnId).includes(option.value)}
                 <button
                   type="button"
                   class="flex min-h-12 w-full items-center gap-3 rounded-lg px-3 text-left active:bg-muted"
-                  onclick={() => toggleValue(screen.columnId, option.value)}
+                  onclick={() => toggleValue(columnId, option.value)}
                 >
                   <span
                     class={cn(
@@ -1010,122 +698,7 @@
           </div>
         {/if}
       {:else if screen.kind === 'date'}
-        <div class="max-h-[72dvh] overflow-y-auto p-4">
-          <div class="mb-4 flex gap-2 overflow-x-auto">
-            {#each [{ value: 'single', label: 'Exact' }, { value: 'range', label: 'Range' }, { value: 'from', label: 'From' }, { value: 'to', label: 'Until' }] as mode (mode.value)}
-              <button
-                type="button"
-                class={cn(
-                  'h-8 shrink-0 rounded-full border px-3 text-xs font-semibold transition-colors',
-                  dateMode() === mode.value
-                    ? 'border-foreground bg-foreground text-background'
-                    : 'bg-background text-muted-foreground active:bg-muted',
-                )}
-                onclick={() =>
-                  setDateMode(mode.value as 'single' | 'range' | 'from' | 'to')}
-              >
-                {mode.label}
-              </button>
-            {/each}
-          </div>
-
-          <DateRangePicker.Root
-            value={dateRangeValue}
-            onValueChange={handleDateRangeChange}
-            bind:placeholder={calendarPlaceholder}
-            weekdayFormat="short"
-            fixedWeeks
-            closeOnRangeSelect={false}
-            class="mx-auto flex w-full max-w-[340px] flex-col gap-2"
-          >
-            <DateRangePicker.Calendar
-              class="rounded-xl border bg-background p-3 shadow-xs"
-            >
-              {#snippet children({ months, weekdays })}
-                <DateRangePicker.Header
-                  class="flex items-center justify-between"
-                >
-                  <CalendarParts.PrevButton>
-                    <ChevronLeft class="size-4" />
-                  </CalendarParts.PrevButton>
-                  <CalendarParts.Heading
-                    bind:placeholder={calendarPlaceholder}
-                    bind:mode={calendarMode}
-                  />
-                  <CalendarParts.NextButton>
-                    <ChevronRight class="size-4" />
-                  </CalendarParts.NextButton>
-                </DateRangePicker.Header>
-                <div class="pt-3">
-                  {#each months as month (month.value)}
-                    <DateRangePicker.Grid
-                      class="w-full border-collapse select-none space-y-1"
-                    >
-                      <DateRangePicker.GridHead>
-                        <DateRangePicker.GridRow
-                          class="mb-1 flex w-full justify-between"
-                        >
-                          {#each weekdays as day (day)}
-                            <DateRangePicker.HeadCell
-                              class="w-10 text-center text-xs font-normal text-muted-foreground"
-                            >
-                              {day.slice(0, 2)}
-                            </DateRangePicker.HeadCell>
-                          {/each}
-                        </DateRangePicker.GridRow>
-                      </DateRangePicker.GridHead>
-                      <DateRangePicker.GridBody>
-                        {#each month.weeks as weekDates (weekDates)}
-                          <DateRangePicker.GridRow class="flex w-full">
-                            {#each weekDates as date (date)}
-                              <DateRangePicker.Cell
-                                {date}
-                                month={month.value}
-                                class="relative m-0 size-10 overflow-visible p-0 text-center text-sm"
-                              >
-                                <DateRangePicker.Day
-                                  class={cn(
-                                    'group relative inline-flex size-10 items-center justify-center rounded-md border border-transparent bg-transparent p-0 text-sm font-normal transition-colors',
-                                    'hover:bg-muted focus-visible:ring-1 focus-visible:ring-ring',
-                                    'data-disabled:pointer-events-none data-disabled:text-foreground/30 data-outside-month:pointer-events-none data-outside-month:text-muted-foreground/40',
-                                    'data-highlighted:rounded-none data-highlighted:bg-muted',
-                                    'data-selected:bg-muted data-selected:[&:not([data-selection-start])]:[&:not([data-selection-end])]:rounded-none',
-                                    'data-selection-start:rounded-md data-selection-start:bg-foreground data-selection-start:text-background data-selection-start:font-semibold',
-                                    'data-selection-end:rounded-md data-selection-end:bg-foreground data-selection-end:text-background data-selection-end:font-semibold',
-                                  )}
-                                >
-                                  {date.day}
-                                </DateRangePicker.Day>
-                              </DateRangePicker.Cell>
-                            {/each}
-                          </DateRangePicker.GridRow>
-                        {/each}
-                      </DateRangePicker.GridBody>
-                    </DateRangePicker.Grid>
-                  {/each}
-                </div>
-              {/snippet}
-            </DateRangePicker.Calendar>
-          </DateRangePicker.Root>
-
-          <div class="mt-4 grid grid-cols-2 gap-2">
-            <Button variant="outline" class="h-11" onclick={setThisYear}>
-              This year
-            </Button>
-            <Button
-              variant="outline"
-              class="h-11"
-              onclick={() =>
-                cloneFilters({
-                  years: [],
-                  fromDate: undefined,
-                  toDate: undefined,
-                })}
-            >
-              Clear date
-            </Button>
-          </div>
-        </div>
+        <MobileDateFilterScreen bind:filters />
       {/if}
     </AnimatedSizeContainer>
   </ModalBody>
@@ -1144,14 +717,15 @@
       </Button>
     </ModalFooter>
   {:else if screen.kind === 'filter'}
+    {@const columnId = screen.columnId}
     <ModalFooter class="border-t p-3">
       <Button class="h-11 w-full" onclick={() => go({ kind: 'home' })}>
         Done
-        {#if optionCount(screen.columnId)}
+        {#if optionCount(columnId)}
           <span
             class="ml-1 rounded bg-primary-foreground/15 px-1.5 py-0.5 text-xs"
           >
-            {optionCount(screen.columnId)}
+            {optionCount(columnId)}
           </span>
         {/if}
       </Button>
