+++ src/lib/components/map/Map.svelte
@@ -28,7 +28,10 @@
   import Filters from '$lib/components/flight-filters/Filters.svelte';
   import MobileFiltersModal from '$lib/components/flight-filters/MobileFiltersModal.svelte';
   import {
-    defaultFilters,
+    createDefaultFilters,
+    hasFlightFilters,
+  } from '$lib/components/flight-filters/model';
+  import {
     hasTempFilters as hasActiveTempFilters,
     type FlightFilters,
     type Route,
@@ -107,16 +110,17 @@
   let showPreviousView = $state(false);
   let programmaticCameraMove = false;
   let handledFocusRequest = $state(-1);
+  const currentTheme = $derived(mode.current ?? 'light');
   const style = $derived(
     getConfiguredAppMapStyleUrl(
-      mode.current,
+      currentTheme,
       appConfig.config?.map,
       mapPreferences.basemap,
     ),
   );
-  const images = $derived(getAppMapImages(base, mode.current));
+  const images = $derived(getAppMapImages(base, currentTheme));
   const openAipTheme = $derived(
-    (mode.current === 'dark' ? 'dark' : 'light') as OpenAipTheme,
+    (currentTheme === 'dark' ? 'dark' : 'light') as OpenAipTheme,
   );
   const openAipConfigured = $derived(
     !!appConfig.configured?.integrations?.openAipKey,
@@ -128,7 +132,7 @@
     getOpenAipOverlayLayers(mapPreferences.openAipGroups, openAipTheme),
   );
   const usingDefaultAppStyle = $derived(
-    style === getDefaultAppMapStyleUrl(mode.current, mapPreferences.basemap),
+    style === getDefaultAppMapStyleUrl(currentTheme, mapPreferences.basemap),
   );
   const openAipTileUrlTemplate = $derived(
     browser
@@ -247,22 +251,7 @@
     showPreviousView = false;
   };
 
-  const showClear = $derived.by(() => {
-    return (
-      filters &&
-      (filters.departureAirports.length ||
-        filters.arrivalAirports.length ||
-        filters.airportsEither.length ||
-        filters.routes.length ||
-        filters.years.length ||
-        filters.fromDate ||
-        filters.toDate ||
-        filters.passengers.length ||
-        filters.airline.length ||
-        filters.aircraft.length ||
-        filters.aircraftRegs.length)
-    );
-  });
+  const showClear = $derived(filters ? hasFlightFilters(filters) : false);
 
   const routeMatches = (
     item: { from: { id: number }; to: { id: number } },
@@ -508,7 +497,9 @@
             return t;
           });
           if (!target) return;
-          targetCenter = [target.center.lng, target.center.lat];
+          if (!target.center || target.zoom === undefined) return;
+          const center = maplibregl.LngLat.convert(target.center);
+          targetCenter = [center.lng, center.lat];
           targetZoom = target.zoom;
         }
       }
@@ -662,7 +653,7 @@
           >
             <ControlButton
               onclick={() => {
-                filters = defaultFilters;
+                filters = createDefaultFilters();
               }}
               title="Clear filters"
             >
