+++ src/lib/components/ui/modal/ModalFooter.svelte
@@ -2,15 +2,25 @@
   import type { Snippet } from 'svelte';
 
   import { getModalContext } from './Modal.svelte';
+  import { cn } from '$lib/utils';
 
-  let { children }: { children: Snippet } = $props();
+  let {
+    children,
+    class: className,
+  }: {
+    children: Snippet;
+    class?: string;
+  } = $props();
 
   const ctx = getModalContext();
   ctx.registerFooter();
 </script>
 
 <div
-  class="flex items-center justify-end gap-2 border-t bg-neutral-50 dark:bg-input/30 p-4"
+  class={cn(
+    'flex items-center justify-end gap-2 border-t bg-neutral-50 dark:bg-input/30 p-4',
+    className,
+  )}
 >
   {@render children()}
 </div>
